removes the Terrible Terror from the mail stall in 2989 3043 missions
Adds a setup script that removes the terrible terror perched on the stall at the School for Mission 2989 (Age of the Night Lights) and Mission 3043 task 6547 (Friendship Festival 2022).

For Friendship Festival 2022, the removal of the terrible terror is intended by the developers themselves, as evidenced by an ObMissionActivate monobehavior located in pathID 14489 in HubSchoolDO. I've tried the quest, and yet the terrible terror is still there, phasing through Harold's Terrible Terror during a cutscene. This pull request fixes that.

For Interception, a quest in Age of the Night Lights, users are getting confused when asked to click on Leopold. They think they need to click on the Terrible Terror perched on the stall, when instead they need to click on Leopold, Harald's Terrible Terror, that is flying in the air. This pull request removes the terror to help alleviate the confusion.